improve code readability inside AbstractAuctionGui by splitting buttons
