+++ spigot/src/main/kotlin/ru/astrainteractive/astramarket/command/auction/AuctionCommandExecutor.kt
@@ -6,7 +6,7 @@ import ru.astrainteractive.astralibs.command.api.executor.CommandExecutor
 import ru.astrainteractive.astramarket.api.market.model.MarketSlot
 import ru.astrainteractive.astramarket.command.auction.di.AuctionCommandDependencies
 import ru.astrainteractive.astramarket.domain.usecase.CreateAuctionUseCase
-import ru.astrainteractive.astramarket.presentation.router.GuiRouter
+import ru.astrainteractive.astramarket.gui.router.GuiRouter
 import kotlin.math.max
 import kotlin.math.min
