+++ spigot/src/main/kotlin/ru/astrainteractive/astramarket/di/RootModule.kt
@@ -3,7 +3,7 @@ package ru.astrainteractive.astramarket.di
 import ru.astrainteractive.astramarket.command.di.CommandModule
 import ru.astrainteractive.astramarket.core.di.CoreModule
 import ru.astrainteractive.astramarket.domain.di.SharedDomainModule
-import ru.astrainteractive.astramarket.presentation.di.AuctionGuiModule
+import ru.astrainteractive.astramarket.gui.di.AuctionGuiModule
 
 interface RootModule {
     val coreModule: CoreModule
