+++ spigot/src/main/kotlin/ru/astrainteractive/astramarket/gui/button/BorderButtonFactory.kt
@@ -1,0 +1,21 @@
+package ru.astrainteractive.astramarket.gui.button
+
+import ru.astrainteractive.astralibs.kyori.KyoriComponentSerializer
+import ru.astrainteractive.astralibs.menu.slot.InventorySlot
+import ru.astrainteractive.astralibs.menu.slot.util.InventorySlotBuilderExt.setDisplayName
+import ru.astrainteractive.astralibs.menu.slot.util.InventorySlotBuilderExt.setIndex
+import ru.astrainteractive.astralibs.menu.slot.util.InventorySlotBuilderExt.setItemStack
+import ru.astrainteractive.astramarket.gui.button.di.ButtonFactoryDependency
+import ru.astrainteractive.astramarket.gui.util.ItemStackExt.toItemStack
+
+class BorderButtonFactory(
+    dependency: ButtonFactoryDependency
+) : ButtonFactoryDependency by dependency,
+    KyoriComponentSerializer by dependency.kyoriComponentSerializer {
+
+    fun render(index: Int) = InventorySlot.Builder()
+        .setIndex(index)
+        .setItemStack(config.buttons.border.toItemStack())
+        .setDisplayName(" ")
+        .build()
+}
