+++ spigot/src/main/kotlin/ru/astrainteractive/astramarket/gui/invmap/InventoryMap.kt
@@ -1,4 +1,4 @@
-package ru.astrainteractive.astramarket.presentation.invmap
+package ru.astrainteractive.astramarket.gui.invmap
 
 interface InventoryMap<T> {
     val map: Array<Array<T>>
