+++ spigot/src/main/kotlin/ru/astrainteractive/astramarket/gui/router/GuiRouter.kt
@@ -1,4 +1,4 @@
-package ru.astrainteractive.astramarket.presentation.router
+package ru.astrainteractive.astramarket.gui.router
 
 import org.bukkit.entity.Player
