+++ spigot/src/main/kotlin/ru/astrainteractive/astramarket/gui/router/di/factory/AuctionGuiFactory.kt
@@ -1,16 +1,16 @@
-package ru.astrainteractive.astramarket.presentation.router.di.factory
+package ru.astrainteractive.astramarket.gui.router.di.factory
 
 import org.bukkit.entity.Player
 import ru.astrainteractive.astramarket.core.di.CoreModule
 import ru.astrainteractive.astramarket.di.ApiMarketModule
 import ru.astrainteractive.astramarket.di.BukkitCoreModule
 import ru.astrainteractive.astramarket.domain.di.SharedDomainModule
-import ru.astrainteractive.astramarket.presentation.auctions.AuctionGui
-import ru.astrainteractive.astramarket.presentation.base.AbstractAuctionGui
-import ru.astrainteractive.astramarket.presentation.base.di.AuctionGuiDependencies
+import ru.astrainteractive.astramarket.gui.base.AbstractAuctionGui
+import ru.astrainteractive.astramarket.gui.base.di.AuctionGuiDependencies
+import ru.astrainteractive.astramarket.gui.button.di.MenuDrawerContext
+import ru.astrainteractive.astramarket.gui.expired.ExpiredAuctionGui
+import ru.astrainteractive.astramarket.gui.router.GuiRouter
 import ru.astrainteractive.astramarket.presentation.di.factory.AuctionComponentFactory
-import ru.astrainteractive.astramarket.presentation.expired.ExpiredAuctionGui
-import ru.astrainteractive.astramarket.presentation.router.GuiRouter
 import ru.astrainteractive.klibs.kdi.Factory
 
 @Suppress("LongParameterList")
@@ -38,17 +38,24 @@ class AuctionGuiFactory(
             bukkitCoreModule = bukkitCoreModule,
             router = router
         )
+        val menuDrawerContext = MenuDrawerContext.Default(
+            coreModule = coreModule,
+            sharedDomainModule = sharedDomainModule,
+            bukkitCoreModule = bukkitCoreModule,
+        )
         return if (isExpired) {
             ExpiredAuctionGui(
                 player = player,
                 dependencies = dependencies,
-                auctionComponent = auctionComponent
+                auctionComponent = auctionComponent,
+                menuDrawerContext = menuDrawerContext
             )
         } else {
-            AuctionGui(
+            ru.astrainteractive.astramarket.gui.auctions.AuctionGui(
                 player = player,
                 auctionComponent = auctionComponent,
                 dependencies = dependencies,
+                menuDrawerContext = menuDrawerContext
             )
         }
     }
